% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAI.R
\name{get.s}
\alias{get.s}
\title{Correlation between tAI and Nc adjusted}
\usage{
get.s(tAI, nc, gc3)
}
\arguments{
\item{tAI}{a vector of length n with tAI values for genes}

\item{nc}{a vector of length n with Nc values for genes}

\item{gc3}{a vector of length n with GC content at third codon positions for genes}
}
\description{
Calculates the correlation between tAI and Nc (adjusted for GC content at
third codon positions).
}
\author{
Mario dos Reis
}
