% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAI.R
\name{get.ws}
\alias{get.ws}
\title{Relative adaptiveness values}
\usage{
get.ws(tRNA, s = NULL, sking)
}
\description{
Relative adaptiveness values
}
\examples{
eco.ws <- get.ws(tRNA=ecolik12$trna, sking=1)

}
\author{
Mario dos Reis
}

