% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAI.R
\name{ts.test}
\alias{ts.test}
\title{Monte Carlo test of correlation between tAI and Nc adjusted}
\usage{
ts.test(m, ws, nc, gc3s, ts.obs, samp.size = 500, n = 1000)
}
\description{
Monte Carlo test of correlation between tAI and Nc adjusted
}

